Сделал возвращение id после создания записи и заменил int на long в id

+++ feature/notedetail/src/main/java/com/example/notedetail/presentation/NoteDetailViewModel.kt
@@ -50,7 +50,7 @@ class NoteDetailViewModel(
 			with(currentState) {
 				updateNoteUseCase(
 					Note(
-						id = noteId.toInt(),
+						id = noteId.toLong(),
 						title = title,
 						description = description,
 					)
@@ -65,7 +65,7 @@ class NoteDetailViewModel(
 			with(currentState) {
 				deleteNoteUseCase(
 					Note(
-						id = noteId.toInt(),
+						id = noteId.toLong(),
 						title = title,
 						description = description,
 					)
